requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
html5lib==1.1

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8

# Image processing and OCR
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data processing
pandas==2.1.3
numpy==1.26.2

# Vector database (for RAG)
chromadb==0.4.18

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
loguru==0.7.2

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# LLM Integration
google-generativeai==0.3.2

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

